+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-11-18
+- First official NPM version
+
+### Added
+- autoload during stackload include via get parameters
+- support npm semver package naming and autoload from jsdelivr 
+
+### Fixed
+- error when type is not guessable
+
 ## [1.0.6] - 2017-10-29
 ### Changed
 - License from GPLv3 to MIT